+++ tests/QueryTest.php
@@ -263,8 +263,8 @@ class QueryTest extends TestCase
             'after_match' => ']',
         ];
 
-        $query = new Query();
-        $rows = $query->from('yii2_test_article_index')
+        $rows = (new Query())
+            ->from('yii2_test_article_index')
             ->match($match)
             ->snippetCallback($snippetCallback)
             ->snippetOptions($snippetOptions)
@@ -276,8 +276,8 @@ class QueryTest extends TestCase
         }
 
         // @see https://github.com/yiisoft/yii2-sphinx/issues/61
-        $query = new Query();
-        $rows = $query->from('yii2_test_article_index')
+        $rows = (new Query())
+            ->from('yii2_test_article_index')
             ->match(new Expression(':match', ['match' => '@(content) ' . $connection->escapeMatchValue('about')]))
             ->snippetCallback($snippetCallback)
             ->snippetOptions($snippetOptions)
@@ -287,6 +287,21 @@ class QueryTest extends TestCase
             $this->assertContains($snippetPrefix, $row['snippet'], 'Snippet source not present!');
             $this->assertContains($snippetOptions['before_match'] . $match, $row['snippet'] . $snippetOptions['after_match'], 'Options not applied!');
         }
+
+        // @see https://github.com/yiisoft/yii2-sphinx/pull/93
+        $rows = (new Query())
+            ->from('yii2_test_article_index')
+            ->match($match)
+            ->snippetCallback(function ($rows) {
+                return [
+                    123,
+                    null,
+                ];
+            })
+            ->all($connection);
+        $this->assertNotEmpty($rows);
+        $this->assertSame('123', $rows[0]['snippet']);
+        $this->assertSame('', $rows[1]['snippet']);
     }
 
     public function testCount()